Rename the library and public header to meshlink-tiny.
This allows the library to coexist with the full MeshLink library in the
same installation directory, and avoids accidental linking to the wrong
version of the library.

The function names have not changed, so it is not possible to
simultaneously link both versions of the library into a single program.