Update the test suite for Meshlink-tiny.
It depends on the full Meshlink library being available as well, which must
be installed or its location should be added to LD_LIBRARY_PATH in order to
be able to run the tests. It is being loaded using dlopen() with the
RTLD_LOCAL and RTLD_DEEPBIND flags to ensure the symbols from the tiny and
full version don't collide.